requirements.txt
Pillow~=8.1.0
tkinter
keras~=2.4.3
tensorflow
numpy~=1.19.5